requirements.txt
streamlit
google-generativeai
beautifulsoup4
requests
pandas
python-dotenv